local_env/mlflow/requirements.txt
pandas==2.3.2
numpy==2.0.2
sqlalchemy==1.4.54
psycopg2-binary==2.9.8
python-dotenv==1.1.1
scikit-learn==1.6.1
mlflow==3.1.4
mlflow-skinny==3.1.4
scikit-surprise==1.1.4